+++ src/main/java/duke/Command/MarkCommand.java
@@ -1,3 +1,8 @@
+package duke.Command;
+
+import duke.*;
+import duke.Tasks.Task;
+import duke.Tasks.TaskList;
 
 public class MarkCommand extends Command {
     private int index;
@@ -9,7 +14,7 @@ public class MarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (index < 1 || index > tasks.size()) {
-            throw new DukeException("OOPS!!! Task index is out of range.");
+            throw new DukeException("OOPS!!! duke.Tasks.Task index is out of range.");
         }
 
         Task task = tasks.get(index - 1);
