+++ src/main/java/duke/Command/UnmarkCommand.java
@@ -1,3 +1,9 @@
+package duke.Command;
+
+import duke.*;
+import duke.Tasks.Task;
+import duke.Tasks.TaskList;
+
 public class UnmarkCommand extends Command {
     private int index;
 
@@ -8,7 +14,7 @@ public class UnmarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (index < 1 || index > tasks.size()) {
-            throw new DukeException("OOPS!!! Task index is out of range.");
+            throw new DukeException("OOPS!!! duke.Tasks.Task index is out of range.");
         }
 
         Task task = tasks.get(index - 1);
