+++ src/main/java/duke/Parser.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.Command.*;
+
+
 import java.util.Arrays;
 public class Parser {
     public static Command parse(String input) throws DukeException {
