+++ src/main/java/duke/Tasks/DeadlineTaskLoad.java
@@ -1,3 +1,7 @@
+package duke.Task;
+
+import duke.Task;
+
 public class DeadlineTaskLoad extends Task {
     private String by;
 
@@ -30,7 +34,7 @@ import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Locale;
 
-public class DeadlineTaskLoad extends Task {
+public class duke.Task.DeadlineTaskLoad extends duke.Task {
     private String by;
 
     int check = 0;
@@ -38,7 +42,7 @@ public class DeadlineTaskLoad extends Task {
     private LocalDateTime dateTime;
 
 
-    public DeadlineTaskLoad(String task, String by) {
+    public duke.Task.DeadlineTaskLoad(String task, String by) {
         super(task);
         this.by = by;
         this.dateTime = parseDateTime(by);
