+++ src/main/java/duke/Ui.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.Tasks.Task;
+import duke.Tasks.TaskList;
+
 import java.util.Scanner;
 public class Ui {
     private Scanner scanner;
